refactor(web): split the 2,100-line create_app() into a web/ package
app.py is now a 99-line factory. Routes live in web/<area>.py, each with a
register(app) function using plain @app.route so endpoint names -- and every
url_for('index') in the templates and tests -- are unchanged. Route parity
with the old file verified (40 routes, identical paths and methods).

- web/dashboard.py, accounts.py, sync.py, settings.py, db_admin.py,
  diagnostics.py, certificates.py, auth.py, template_context.py, autostart.py
- bootstrap.py: schema creation, self-healing column migrations, stuck-sync
  reset and provider registration (previously duplicated between main.py and
  the dev runner).
- expiration.py: annotate_account_expiration / annotate_certificate_expiration
  replace the four hand-rolled expired/critical/warning/safe/at_risk blocks;
  the account detail page now honours Korean Air's per-batch earliest expiry
  the same way the dashboard does.
- Provider metadata moves onto the plugins: the two 30-entry domain and
  homepage tables in the context processor become plugin.logo_domain and
  plugin.homepage_url. ProviderPlugin gains is_manual (replaces the
  hard-coded MANUAL_PLUGIN_IDS set) and normalize_username (Korean Air's
  SKYPASS space-stripping moves out of the account form handlers).
- Settings reads/writes go through services.settings_store instead of 36
  inline Settings.query.filter_by(key=...) calls.
