+++ plugins/alaska.py
@@ -16,6 +16,14 @@ class AlaskaAirlinesPlugin(ProviderPlugin):
         return "alaska"
 
     @property
+    def homepage_url(self) -> str:
+        return "https://www.alaskaair.com/mileageplan"
+
+    @property
+    def logo_domain(self) -> str:
+        return "alaskaair.com"
+
+    @property
     def default_cpp(self) -> float:
         return 1.4
