+++ plugins/southwest.py
@@ -16,6 +16,14 @@ class SouthwestPlugin(ProviderPlugin):
         return "southwest"
 
     @property
+    def homepage_url(self) -> str:
+        return "https://www.southwest.com/rapidrewards/"
+
+    @property
+    def logo_domain(self) -> str:
+        return "southwest.com"
+
+    @property
     def default_cpp(self) -> float:
         return 1.3
