+++ plugins/virgin.py
@@ -19,6 +19,14 @@ class VirginAtlanticPlugin(ProviderPlugin):
         return "virgin"
 
     @property
+    def homepage_url(self) -> str:
+        return "https://flywith.virginatlantic.com/gb/en/flying-club.html"
+
+    @property
+    def logo_domain(self) -> str:
+        return "virginatlantic.com"
+
+    @property
     def default_cpp(self) -> float:
         return 1.2
